add support for recognizing locktimes
the inputs are now written to db with a uint64 appended, which denoted the height at which this input unlocks. inputs that are unlocked have this number set to zero. now, only unlocked inputs can be included in a transaction. locked status is checked and updated each time a new block arrives